requirement.txt
python-dotenv
transformer
scikit-learn
scipy
seaborn
torch
matplotlib
numpy
pandasf
flask